requirements.txt
openai>=1.3.0
chromadb>=0.4.18
tiktoken>=0.5.1
numpy>=2.0.0
pandas>=2.0.0
datasets>=2.14.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pysbd>=0.3.4
requests>=2.31.0
certifi>=2024.0.0

# Совместимые версии для Python 3.12 и RAGAS
pydantic>=2.0.0,<3.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
ragas>=0.2.0

